update axis from manager
