+++ bodo/pandas/lazy_wrapper.py
@@ -2,6 +2,7 @@ import abc
 import typing as pt
 from collections.abc import Callable
 
+from bodo.ext import plan_optimizer
 from bodo.pandas.lazy_metadata import LazyMetadata
 
 
@@ -27,3 +28,19 @@ class BodoLazyWrapper(abc.ABC):
     @property
     def _lazy(self) -> bool:
         return self._get_result_id() is not None
+
+    @property
+    def _plan(self):
+        if hasattr(self._mgr, "_plan"):
+            if self._mgr._plan is not None:
+                return self._mgr._plan
+            else:
+                return plan_optimizer.LogicalGetSeriesRead(self._mgr._md_result_id)
+
+        raise NotImplementedError(
+            "Plan not available for this manager, recreate this series with from_pandas"
+        )
+
+    def is_lazy_plan(self):
+        """Returns whether the BodoDataFrame is represented by a plan."""
+        return getattr(self._mgr, "_plan", None) is not None
