+++ bodo/pandas/series.py
@@ -23,18 +23,6 @@ class BodoSeries(pd.Series, BodoLazyWrapper):
     _head_s: pd.Series | None = None
     _name: Hashable = None
 
-    @property
-    def _plan(self):
-        if hasattr(self._mgr, "_plan"):
-            if self._mgr._plan is not None:
-                return self._mgr._plan
-            else:
-                return plan_optimizer.LogicalGetSeriesRead(self._mgr._md_result_id)
-
-        raise NotImplementedError(
-            "Plan not available for this manager, recreate this series with from_pandas"
-        )
-
     @check_args_fallback("all")
     def _cmp_method(self, other, op):
         """Called when a BodoSeries is compared with a different entity (other)
@@ -130,6 +118,17 @@ class BodoSeries(pd.Series, BodoLazyWrapper):
         else:
             # If head_s is available and larger than n, then use it directly.
             return self._head_s.head(n)
+
+    def __len__(self):
+        if self.is_lazy_plan():
+            self._mgr._collect()
+        return super().__len__()
+
+    @property
+    def shape(self):
+        if self.is_lazy_plan():
+            self._mgr._collect()
+        return super().shape
 
     def _get_result_id(self) -> str | None:
         if isinstance(self._mgr, LazyMetadataMixin):
