+++ bodo/tests/test_df_lib/test_end_to_end.py
@@ -79,4 +79,6 @@ def test_projection(datapath):
     py_df1 = pd.read_parquet(datapath("dataframe_library/df1.parquet"))
     py_df2 = py_df1["D"]
 
+    print(type(bodo_df2))
+
     _test_equal(bodo_df2, py_df2, check_pandas_types=False)
